refactor: ♻️ Refactor pivot operations caching logic into trait.
